Moved ResourcePackLoader as a separated class
